Refactoring - Added db folder
