+++ Cookbook.Infrastructur/Repositories/RecipebookRepository.cs
@@ -1,5 +1,6 @@
 ﻿using Cookbook.Application.Database;
 using Cookbook.Domain.Models;
+using Cookbook.Repository.Database;
 using Dapper;
 using System.Data;
